+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.19] - 2026-04-05
+### Fixed
+- Repository names in multi-repo projects now use simple directory names if they are unambiguous (Github #10 part 3)
 
 ## [0.5.18] - 2026-04-05